+++ geotrek/trekking/views.py
@@ -23,9 +23,9 @@ from mapentity.views import (MapEntityLayer, MapEntityList, MapEntityJsonList,
 from rest_framework import permissions as rest_permissions, viewsets
 
 from geotrek.authent.decorators import same_structure_required
-from geotrek.common.mixins import transform_pdf_booklet_callback
 from geotrek.common.models import Attachment, RecordSource, TargetPortal
-from geotrek.common.views import FormsetMixin, MetaMixin, PublicOrReadPermMixin, DocumentPublic, MarkupPublic
+from geotrek.common.views import (FormsetMixin, MetaMixin, PublicOrReadPermMixin, DocumentPublic,
+                                  DocumentBookletPublic, MarkupPublic)
 from geotrek.core.models import AltimetryMixin
 from geotrek.core.views import CreateFromTopologyMixin
 from geotrek.trekking.forms import SyncRandoForm
@@ -45,8 +45,6 @@ from geotrek.infrastructure.serializers import InfrastructureGeojsonSerializer
 from geotrek.signage.serializers import SignageGeojsonSerializer
 
 from .tasks import launch_sync_rando
-
-from mapentity.helpers import suffix_for, smart_get_template
 
 
 class SyncRandoRedirect(RedirectView):
@@ -217,17 +215,11 @@ class TrekDocumentPublicMixin(object):
 
 
 class TrekDocumentPublic(TrekDocumentPublicMixin, DocumentPublic):
-    def __init__(self, *args, **kwargs):
-        super(TrekDocumentPublic, self).__init__(*args, **kwargs)
-        if settings.USE_BOOKLET_PDF:
-            suffix = suffix_for(self.template_name_suffix, "_booklet_pdf", "html")
-            self.template_name = smart_get_template(self.model, suffix)
+    pass
 
-    def get(self, request, pk, slug, lang=None):
-        response = super(TrekDocumentPublic, self).get(request, pk, slug)
-        if settings.USE_BOOKLET_PDF:
-            response.add_post_render_callback(transform_pdf_booklet_callback)
-        return response
+
+class TrekBookletDocumentPublic(TrekDocumentPublicMixin, DocumentBookletPublic):
+    pass
 
 
 class TrekMarkupPublic(TrekDocumentPublicMixin, MarkupPublic):
